I removed the   #' @importFrom latex2exp From Namespace. I tnow onyl has GGplot2 in there.
I then changed how the plot_linreg.R function works so it does not depend on    latex2exp anymore. It does not come out as well but i think we had problems with dependencies on this small indie package! Better to go with a well established Package only like ggplot!